backend/pyproject.toml
[project]
name = "furniture"
version = "0.1.0"
description = ""
authors = [
    {name = "Kirill"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
